Unifying logic for invoice downloading
